requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
babel
backoff
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
courlan
cryptography
dataclasses-json
dateparser
deepdiff
distro
docstring_parser
emoji
faiss-cpu
fast-graphrag
fastapi
filelock
filetype
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
groq
grpcio
grpcio-status
h11
hnswlib
html5lib
htmldate
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
idna
igraph
instructor
InstructorEmbedding
Jinja2
jiter
joblib
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
jusText
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-openai
langchain-text-splitters
langdetect
langsmith
llm-components
lxml
lxml_html_clean
markdown-it-py
markdownify
MarkupSafe
marshmallow
mdurl
mpmath
multidict
mypy-extensions
narwhals
nest-asyncio
networkx
nltk
numpy
olefile
openai
optype
orderly-set
orjson
packaging
pandas
pathspec
pdfminer.six
pdfplumber
pillow
pinecone
pinecone-plugin-inference
pinecone-plugin-interface
propcache
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyMuPDF
pyparsing
pypdf
PyPDF2
pypdfium2
python-dateutil
python-dotenv
python-iso639
python-magic
python-magic-bin
python-multipart
python-oxmsg
pytz
PyYAML
RapidFuzz
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
scipy-stubs
sentence-transformers
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
streamlit
structlog
sympy
tenacity
texttable
threadpoolctl
tiktoken
tld
tokenizers
toml
torch
tornado
tqdm
trafilatura
transformers
typer
typing-inspect
typing_extensions
tzdata
tzlocal
unstructured
unstructured-client
uritemplate
urllib3
uvicorn
watchdog
watchfiles
webencodings
websockets
wrapt
xxhash
yarl
lightrag-hku
aioboto3
ollama
nano_vectordb
sentencepiece
accelerate